// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};
// ...
export type GeneralSettings = {
  __typename?: 'GeneralSettings';
  /** A date format for all date strings. */
  dateFormat?: Maybe<Scalars['String']>;
  /** Site tagline. */
  description?: Maybe<Scalars['String']>;
  /** This address is used for admin purposes, like new user notification. */
  email?: Maybe<Scalars['String']>;
  /** WordPress locale code. */
  language?: Maybe<Scalars['String']>;
  /** A day number of the week that the week should start on. */
  startOfWeek?: Maybe<Scalars['Int']>;
  /** A time format for all time strings. */
  timeFormat?: Maybe<Scalars['String']>;
  /** A city in the same timezone as you. */
  timezone?: Maybe<Scalars['String']>;
  /** Site title. */
  title?: Maybe<Scalars['String']>;
  /** Site URL. */
  url?: Maybe<Scalars['String']>;
};
// ...
export type MenuItemToMenuConnectionEdge = Edge & MenuConnectionEdge & OneToOneConnection & {
  __typename?: 'MenuItemToMenuConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']>;
  /** The node of the connection, without the edges */
  node: Menu;
};
// ...
export type NodeWithAuthorToUserConnectionEdge = Edge & OneToOneConnection & UserConnectionEdge & {
  __typename?: 'NodeWithAuthorToUserConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']>;
  /** The node of the connection, without the edges */
  node: User;
};
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Page_Home = AcfFieldGroup & {
  __typename?: 'Page_Home';
  blogSection?: Maybe<Page_Home_BlogSection>;
  contactSection?: Maybe<Page_Home_ContactSection>;
  ctaSection?: Maybe<Page_Home_CtaSection>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heroSection?: Maybe<Page_Home_HeroSection>;
  introSection?: Maybe<Page_Home_IntroSection>;
  testimonialsSection?: Maybe<Page_Home_TestimonialsSection>;
};

/** Field Group */
export type Page_Home_BlogSection = AcfFieldGroup & {
  __typename?: 'Page_Home_BlogSection';
  body?: Maybe<Scalars['String']>;
  featuredPosts?: Maybe<Array<Maybe<PostObjectUnion>>>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heading?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Home_ContactSection = AcfFieldGroup & {
  __typename?: 'Page_Home_ContactSection';
  emailAddress?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  phoneNumber?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Home_CtaSection = AcfFieldGroup & {
  __typename?: 'Page_Home_CtaSection';
  body?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heading?: Maybe<Scalars['String']>;
  link?: Maybe<PostObjectUnion>;
};

/** Field Group */
export type Page_Home_HeroSection = AcfFieldGroup & {
  __typename?: 'Page_Home_HeroSection';
  backgroundImage?: Maybe<MediaItem>;
  body?: Maybe<Scalars['String']>;
  cta?: Maybe<PostObjectUnion>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  headingRepeater?: Maybe<Array<Maybe<Page_Home_HeroSection_HeadingRepeater>>>;
  youtubeLink?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Home_HeroSection_HeadingRepeater = AcfFieldGroup & {
  __typename?: 'Page_Home_HeroSection_headingRepeater';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  isAccented?: Maybe<Scalars['Boolean']>;
  text?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Home_IntroSection = AcfFieldGroup & {
  __typename?: 'Page_Home_IntroSection';
  body?: Maybe<Scalars['String']>;
  featuredList?: Maybe<Array<Maybe<Page_Home_IntroSection_FeaturedList>>>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heading?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Page_Home_IntroSection_FeaturedList = AcfFieldGroup & {
  __typename?: 'Page_Home_IntroSection_featuredList';
  body?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heading?: Maybe<Scalars['String']>;
  image?: Maybe<MediaItem>;
};

/** Field Group */
export type Page_Home_TestimonialsSection = AcfFieldGroup & {
  __typename?: 'Page_Home_TestimonialsSection';
  body?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  heading?: Maybe<Scalars['String']>;
  image?: Maybe<MediaItem>;
  testimonialList?: Maybe<Array<Maybe<Page_Home_TestimonialsSection_TestimonialList>>>;
};
// ...
export type PostObjectUnion = MediaItem | Page | Post;
// ...
export type RootQueryToMenuItemConnection = Connection & MenuItemConnection & {
  __typename?: 'RootQueryToMenuItemConnection';
  /** Edges for the RootQueryToMenuItemConnection connection */
  edges: Array<RootQueryToMenuItemConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<MenuItem>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToMenuItemConnectionPageInfo;
};
// ...
export type User = Commenter & DatabaseIdentifier & Node & UniformResourceIdentifiable & {
  __typename?: 'User';
  /** Avatar object for user. The avatar object can be retrieved in different sizes by specifying the size argument. */
  avatar?: Maybe<Avatar>;
  /** User metadata option name. Usually it will be &quot;wp_capabilities&quot;. */
  capKey?: Maybe<Scalars['String']>;
  /** A list of capabilities (permissions) granted to the user */
  capabilities?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Connection between the User type and the Comment type */
  comments?: Maybe<UserToCommentConnection>;
  /** @deprecated Deprecated in favor of using Next.js pages */
  conditionalTags?: Maybe<ConditionalTags>;
  /** Identifies the primary key from the database. */
  databaseId: Scalars['Int'];
  /** Description of the user. */
  description?: Maybe<Scalars['String']>;
  /** Email address of the user. This is equivalent to the WP_User-&gt;user_email property. */
  email?: Maybe<Scalars['String']>;
  /** Connection between the User type and the EnqueuedScript type */
  enqueuedScripts?: Maybe<UserToEnqueuedScriptConnection>;
  /** Connection between the User type and the EnqueuedStylesheet type */
  enqueuedStylesheets?: Maybe<UserToEnqueuedStylesheetConnection>;
  /** A complete list of capabilities including capabilities inherited from a role. This is equivalent to the array keys of WP_User-&gt;allcaps. */
  extraCapabilities?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** First name of the user. This is equivalent to the WP_User-&gt;user_first_name property. */
  firstName?: Maybe<Scalars['String']>;
  /** The globally unique identifier for the user object. */
  id: Scalars['ID'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** Last name of the user. This is equivalent to the WP_User-&gt;user_last_name property. */
  lastName?: Maybe<Scalars['String']>;
  /** The preferred language locale set for the user. Value derived from get_user_locale(). */
  locale?: Maybe<Scalars['String']>;
  /** Connection between the User type and the mediaItem type */
  mediaItems?: Maybe<UserToMediaItemConnection>;
  /** Display name of the user. This is equivalent to the WP_User-&gt;dispaly_name property. */
  name?: Maybe<Scalars['String']>;
  /** The nicename for the user. This field is equivalent to WP_User-&gt;user_nicename */
  nicename?: Maybe<Scalars['String']>;
  /** Nickname of the user. */
  nickname?: Maybe<Scalars['String']>;
  /** Connection between the User type and the page type */
  pages?: Maybe<UserToPageConnection>;
  /** Connection between the User type and the post type */
  posts?: Maybe<UserToPostConnection>;
  /** The date the user registered or was created. The field follows a full ISO8601 date string format. */
  registeredDate?: Maybe<Scalars['String']>;
  /** Connection between the User and Revisions authored by the user */
  revisions?: Maybe<UserToRevisionsConnection>;
  /** Connection between the User type and the UserRole type */
  roles?: Maybe<UserToUserRoleConnection>;
  /** The Yoast SEO data of a user */
  seo?: Maybe<SeoUser>;
  /** Whether the Toolbar should be displayed when the user is viewing the site. */
  shouldShowAdminToolbar?: Maybe<Scalars['Boolean']>;
  shouldShowFaustToolbar?: Maybe<Scalars['Boolean']>;
  /** The slug for the user. This field is equivalent to WP_User-&gt;user_nicename */
  slug?: Maybe<Scalars['String']>;
  templates?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
  /** A website url that is associated with the user. */
  url?: Maybe<Scalars['String']>;
  /**
   * The Id of the user. Equivalent to WP_User-&gt;ID
   * @deprecated Deprecated in favor of the databaseId field
   */
  userId?: Maybe<Scalars['Int']>;
  /** Username for the user. This field is equivalent to WP_User-&gt;user_login. */
  username?: Maybe<Scalars['String']>;
};
// ...
export type GetHomePageQuery = { __typename?: 'RootQuery', page?: { __typename?: 'Page', title?: string | null, content?: string | null, seo?: { __typename?: 'PostTypeSEO', fullHead?: string | null } | null, home?: { __typename?: 'Page_Home', heroSection?: { __typename?: 'Page_Home_HeroSection', body?: string | null, headingRepeater?: Array<{ __typename?: 'Page_Home_HeroSection_headingRepeater', text?: string | null, isAccented?: boolean | null } | null> | null, cta?: { __typename?: 'MediaItem' } | { __typename?: 'Page', title?: string | null, uri?: string | null } | { __typename?: 'Post' } | null } | null, introSection?: { __typename?: 'Page_Home_IntroSection', heading?: string | null, body?: string | null, featuredList?: Array<{ __typename?: 'Page_Home_IntroSection_featuredList', heading?: string | null, body?: string | null, image?: { __typename?: 'MediaItem', sourceUrl?: string | null } | null } | null> | null } | null, testimonialsSection?: { __typename?: 'Page_Home_TestimonialsSection', heading?: string | null, body?: string | null, testimonialList?: Array<{ __typename?: 'Page_Home_TestimonialsSection_testimonialList', name?: string | null, quote?: string | null, location?: string | null } | null> | null } | null, ctaSection?: { __typename?: 'Page_Home_CtaSection', heading?: string | null, body?: string | null, link?: { __typename?: 'MediaItem' } | { __typename?: 'Page', uri?: string | null, title?: string | null } | { __typename?: 'Post' } | null } | null, blogSection?: { __typename?: 'Page_Home_BlogSection', heading?: string | null, body?: string | null, featuredPosts?: Array<{ __typename?: 'MediaItem' } | { __typename?: 'Page' } | { __typename?: 'Post', id: string, title?: string | null, excerpt?: string | null, date?: string | null, uri?: string | null, author?: { __typename?: 'NodeWithAuthorToUserConnectionEdge', node: { __typename?: 'User', name?: string | null } } | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', sourceUrl?: string | null } } | null } | null> | null } | null, contactSection?: { __typename?: 'Page_Home_ContactSection', phoneNumber?: string | null, emailAddress?: string | null } | null } | null } | null, generalSettings?: { __typename?: 'GeneralSettings', title?: string | null, description?: string | null } | null, primaryMenuItems?: { __typename?: 'RootQueryToMenuItemConnection', nodes: Array<{ __typename?: 'MenuItem', id: string, uri?: string | null, path?: string | null, label?: string | null, parentId?: string | null, cssClasses?: Array<string | null> | null, menu?: { __typename?: 'MenuItemToMenuConnectionEdge', node: { __typename?: 'Menu', name?: string | null } } | null }> } | null };